perf(test): cut workflow guard import overhead
